Change login.msonline.com to portal.office.com, correct domain name, typos
While not a problem in the lab, the use of login.microsoftonline.com in a production environment where the user has other Microsoft cloud services, could be a problem. Changed yourdomain.hostdomain to the agreed upon adatumyyxxxxx.hostdomain.com. Typos: a couple of misplaced spaces and added escape character before "*"s in the SPAM test strings.